Separate settings screen from profile screen
