+++ lib/features/auth/screens/login.dart
@@ -2,6 +2,7 @@ import 'package:flextras/flextras.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:gap/gap.dart';
+import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../../../services/api/api_client.dart';
@@ -37,8 +38,8 @@ class LoginScreen extends HookConsumerWidget {
         title: const Text('Login'),
         actions: [
           IconButton(
-            onPressed: () => context.showAppLicensePage(),
-            icon: const Icon(Icons.info_outline),
+            onPressed: () => context.push('/settings'),
+            icon: const Icon(Icons.settings),
           ),
         ],
       ),
