+++ lib/features/profile/screens/profile.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../../services/settings.dart';
-import '../../../utils/extensions.dart';
 import '../../auth/providers/auth_state.dart';
 import '../providers/profile.dart';
 
@@ -12,20 +11,12 @@ class ProfileScreen extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final profile = ref.watch(profileProvider);
-    final themeMode = ref.watch(currentThemeModeProvider);
 
     final profileRecords = [
       (label: 'Name', text: profile?.fullName),
       (label: 'Username', text: profile?.username),
       (label: 'Email', text: profile?.email),
     ];
-
-    void showThemeModeDialog() {
-      showDialog(
-        context: context,
-        builder: (_) => const _ThemeModeDialog(),
-      );
-    }
 
     void logout() {
       ref.read(currentAuthStateProvider.notifier).logout();
@@ -36,8 +27,8 @@ class ProfileScreen extends ConsumerWidget {
         title: const Text('Profile'),
         actions: [
           IconButton(
-            onPressed: () => context.showAppLicensePage(),
-            icon: const Icon(Icons.info_outline),
+            onPressed: () => context.push('/settings'),
+            icon: const Icon(Icons.settings),
           ),
         ],
       ),
@@ -50,57 +41,12 @@ class ProfileScreen extends ConsumerWidget {
             ),
           const Divider(),
           ListTile(
-            leading: const Icon(Icons.brightness_6),
-            title: const Text('Theme mode'),
-            trailing: Text(themeMode.label),
-            onTap: showThemeModeDialog,
-          ),
-          const Divider(),
-          ListTile(
             leading: const Icon(Icons.logout),
             title: const Text('Logout'),
             onTap: logout,
           ),
         ],
       ),
-    );
-  }
-}
-
-class _ThemeModeDialog extends ConsumerWidget {
-  const _ThemeModeDialog();
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    void setThemeMode(ThemeMode themeMode) {
-      ref.read(currentThemeModeProvider.notifier).set(themeMode);
-      Navigator.of(context).pop();
-    }
-
-    return SimpleDialog(
-      clipBehavior: Clip.antiAlias,
-      children: [
-        for (final themeMode in ThemeMode.values)
-          _ThemeModeDialogOption(
-            value: themeMode,
-            onTap: () => setThemeMode(themeMode),
-          )
-      ],
-    );
-  }
-}
-
-class _ThemeModeDialogOption extends StatelessWidget {
-  const _ThemeModeDialogOption({required this.value, required this.onTap});
-
-  final ThemeMode value;
-  final GestureTapCallback? onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      onTap: onTap,
-      title: Text(value.label),
     );
   }
 }
